tweaks

+++ nal/src/main/java/nars/task/TruthPolation.java
@@ -20,9 +20,9 @@ public enum TruthPolation  {
     ;
 
     /** dt > 0 */
-    public static float evidenceDecay(float evi, int dur, float dt) {
+    public static float evidenceDecay(float evi, int dur, long dt) {
 
-        return evi * 1f/( 1f + sqr( dt/dur) ); //inverse square
+        return evi * 1f/( 1f + sqr( ((float)dt)/dur) ); //inverse square
         //return evi * 1f/( 1 + (dt/dur) ); //inverse linear
         //return evi * 1f/( 1 + 2 * (dt/dur) ); //inverse linear * 2
